moved utility color functions out
