+++ src/utilities.js
@@ -1,0 +1,79 @@
+export function rgbToHsl(rgb) {
+	// const [r, g, b] = rgbStr.slice(4, -1).split(',').map(Number);
+	const [r, g, b] = Object.values(rgb);
+	const max = Math.max(r, g, b)
+	const min = Math.min(r, g, b)
+	const l = Math.floor((max + min) / ((0xff * 2) / 100))
+  
+	if (max === min) return [0, 0, l / 100]
+	const d = max - min
+	const s = Math.floor((d / (l > 50 ? 0xff * 2 - max - min : max + min)) * 100)
+  
+	if (max === r) return [Math.floor(((g - b) / d + (g < b && 6)) * 60), s / 100, l / 100]
+	return max === g
+	  ? [Math.floor(((b - r) / d + 2) * 60), s / 100, l / 100]
+	  : [Math.floor(((r - g) / d + 4) * 60), s / 100, l / 100]
+}
+
+/* https://gist.github.com/mjackson/5311256 by kigiri*/
+export function rgbStrToHsl(rgbStr) {
+	const [r, g, b] = rgbStr.slice(4, -1).split(',').map(Number)
+	const max = Math.max(r, g, b)
+	const min = Math.min(r, g, b)
+	const l = Math.floor((max + min) / ((0xff * 2) / 100))
+	
+	if (max === min) return [0, 0, l]
+	const d = max - min
+	const s = Math.floor((d / (l > 50 ? 0xff * 2 - max - min : max + min)) * 100)
+	
+	if (max === r) return [Math.floor(((g - b) / d + (g < b && 6)) * 60), s, l]
+	return max === g
+		? [Math.floor(((b - r) / d + 2) * 60), s, l]
+		: [Math.floor(((r - g) / d + 4) * 60), s, l]
+}
+
+/* https://stackoverflow.com/questions/5623838/rgb-to-hex-and-hex-to-rgb */
+export function hexToRgb(hex) {
+	// Expand shorthand form (e.g. "03F") to full form (e.g. "0033FF")
+	var shorthandRegex = /^#?([a-f\d])([a-f\d])([a-f\d])$/i;
+	hex = hex.replace(shorthandRegex, function(m, r, g, b) {
+		return r + r + g + g + b + b;
+	});
+	
+	var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
+	return result ? {
+		r: parseInt(result[1], 16),
+		g: parseInt(result[2], 16),
+		b: parseInt(result[3], 16)
+	} : null;
+}
+
+export function HSVToRGBString(h,s,v) { //duplicate from Joy.js
+  if(arguments.length===1) {
+        s=h[1], v=h[2], h=h[0]; // cast to different vars
+    }
+  var rgb = HSVtoRGB(h,s,v);
+  return "rgb("+rgb[0]+","+rgb[1]+","+rgb[2]+")";
+}
+
+export function HSVtoRGB(h,s,v) { //duplicate from Joy.js
+	var r, g, b, i, f, p, q, t;
+	if (arguments.length === 1) {
+			s = h.s, v = h.v, h = h.h;
+	}
+	h /= 360; // convert, yo.
+	i = Math.floor(h * 6);
+	f = h * 6 - i;
+	p = v * (1 - s);
+	q = v * (1 - f * s);
+	t = v * (1 - (1 - f) * s);
+	switch (i % 6) {
+			case 0: r = v, g = t, b = p; break;
+			case 1: r = q, g = v, b = p; break;
+			case 2: r = p, g = v, b = t; break;
+			case 3: r = p, g = q, b = v; break;
+			case 4: r = t, g = p, b = v; break;
+			case 5: r = v, g = p, b = q; break;
+	}
+	return [Math.round(r*255), Math.round(g*255), Math.round(b*255)];
+}
